@YashasSamaga 2019-12-16
If anybody is using high-end GPUs and wants to better performance for FP16 can try https://github.com/YashasSamaga/opencv/tree/cuda4dnn-fp-conversion

It appears to hurt performance on low-end GPUs. Hence, I won't be opening a PR until I sort it out.

**Device:** RTX 2080 Ti

Model                      | FP16 (without this PR)  | FP16 (with this PR)
-------------------------- | ----------------------- | -------------------------
ResNet50                   | 2.36ms                  | 2.24ms
YOLOv3*                    | 6.49ms                  | 5.43ms
FNS Stary Night            | 3.04ms                  | 2.91ms
EAST Text Detection        | 3.05ms                  | 2.93ms
Inception v2 Faster RCNN   | 25.68m                  | 16.37ms
Inception v2 Mask RCNN     | 13.51ms                 | 12.23ms

*without NMS in region layer